adk-project/tools/requirements.txt
requests>=2.31.0
ibm-watsonx-orchestrate-adk
